Rebase BeaconBlockBodyMerge on top of BeaconBlockBodyAltair
