+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/blocks/blockbody/versions/altair/BeaconBlockBodySchemaAltairImpl.java
@@ -1,0 +1,155 @@
+/*
+ * Copyright 2021 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.altair;
+
+import java.util.function.Consumer;
+import tech.pegasys.teku.spec.datastructures.blocks.Eth1Data;
+import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBodyBuilder;
+import tech.pegasys.teku.spec.datastructures.blocks.blockbody.common.BlockBodyFields;
+import tech.pegasys.teku.spec.datastructures.operations.Attestation;
+import tech.pegasys.teku.spec.datastructures.operations.AttesterSlashing;
+import tech.pegasys.teku.spec.datastructures.operations.Deposit;
+import tech.pegasys.teku.spec.datastructures.operations.ProposerSlashing;
+import tech.pegasys.teku.spec.datastructures.operations.SignedVoluntaryExit;
+import tech.pegasys.teku.spec.datastructures.type.SszSignature;
+import tech.pegasys.teku.spec.datastructures.type.SszSignatureSchema;
+import tech.pegasys.teku.ssz.SszList;
+import tech.pegasys.teku.ssz.containers.ContainerSchema9;
+import tech.pegasys.teku.ssz.primitive.SszBytes32;
+import tech.pegasys.teku.ssz.schema.SszListSchema;
+import tech.pegasys.teku.ssz.schema.SszPrimitiveSchemas;
+import tech.pegasys.teku.ssz.tree.TreeNode;
+
+class BeaconBlockBodySchemaAltairImpl
+    extends ContainerSchema9<
+        BeaconBlockBodyAltairImpl,
+        SszSignature,
+        Eth1Data,
+        SszBytes32,
+        SszList<ProposerSlashing>,
+        SszList<AttesterSlashing>,
+        SszList<Attestation>,
+        SszList<Deposit>,
+        SszList<SignedVoluntaryExit>,
+        SyncAggregate>
+    implements BeaconBlockBodySchemaAltair<BeaconBlockBodyAltairImpl> {
+
+  private BeaconBlockBodySchemaAltairImpl(
+      NamedSchema<SszSignature> randaoRevealSchema,
+      NamedSchema<Eth1Data> eth1DataSchema,
+      NamedSchema<SszBytes32> graffitiSchema,
+      NamedSchema<SszList<ProposerSlashing>> proposerSlashingsSchema,
+      NamedSchema<SszList<AttesterSlashing>> attesterSlashingsSchema,
+      NamedSchema<SszList<Attestation>> attestationsSchema,
+      NamedSchema<SszList<Deposit>> depositsSchema,
+      NamedSchema<SszList<SignedVoluntaryExit>> voluntaryExitsSchema,
+      NamedSchema<SyncAggregate> syncAggregateSchema) {
+    super(
+        "BeaconBlockBody",
+        randaoRevealSchema,
+        eth1DataSchema,
+        graffitiSchema,
+        proposerSlashingsSchema,
+        attesterSlashingsSchema,
+        attestationsSchema,
+        depositsSchema,
+        voluntaryExitsSchema,
+        syncAggregateSchema);
+  }
+
+  static BeaconBlockBodySchemaAltair<? extends BeaconBlockBodyAltair> create(
+      final long maxProposerSlashings,
+      final long maxAttesterSlashings,
+      final long maxAttestations,
+      final long maxDeposits,
+      final long maxVoluntaryExits,
+      final int syncCommitteeSize) {
+    return new BeaconBlockBodySchemaAltairImpl(
+        namedSchema(BlockBodyFields.RANDAO_REVEAL.name(), SszSignatureSchema.INSTANCE),
+        namedSchema(BlockBodyFields.ETH1_DATA.name(), Eth1Data.SSZ_SCHEMA),
+        namedSchema(BlockBodyFields.GRAFFITI.name(), SszPrimitiveSchemas.BYTES32_SCHEMA),
+        namedSchema(
+            BlockBodyFields.PROPOSER_SLASHINGS.name(),
+            SszListSchema.create(ProposerSlashing.SSZ_SCHEMA, maxProposerSlashings)),
+        namedSchema(
+            BlockBodyFields.ATTESTER_SLASHINGS.name(),
+            SszListSchema.create(AttesterSlashing.SSZ_SCHEMA, maxAttesterSlashings)),
+        namedSchema(
+            BlockBodyFields.ATTESTATIONS.name(),
+            SszListSchema.create(Attestation.SSZ_SCHEMA, maxAttestations)),
+        namedSchema(
+            BlockBodyFields.DEPOSITS.name(), SszListSchema.create(Deposit.SSZ_SCHEMA, maxDeposits)),
+        namedSchema(
+            BlockBodyFields.VOLUNTARY_EXITS.name(),
+            SszListSchema.create(SignedVoluntaryExit.SSZ_SCHEMA, maxVoluntaryExits)),
+        namedSchema(
+            BlockBodyFields.SYNC_AGGREGATE.name(), SyncAggregateSchema.create(syncCommitteeSize)));
+  }
+
+  @Override
+  public BeaconBlockBodyAltairImpl createBlockBody(
+      final Consumer<BeaconBlockBodyBuilder> bodyBuilder) {
+    final BeaconBlockBodyBuilderAltair bodyContent =
+        new BeaconBlockBodyBuilderAltair().schema(this);
+    // Provide a default empty sync aggregate
+    bodyContent.syncAggregate(getSyncAggregateSchema()::createEmpty);
+    bodyBuilder.accept(bodyContent);
+    return bodyContent.build();
+  }
+
+  @Override
+  public BeaconBlockBodyAltairImpl createEmpty() {
+    return new BeaconBlockBodyAltairImpl(this);
+  }
+
+  @SuppressWarnings("unchecked")
+  @Override
+  public SszListSchema<ProposerSlashing, ?> getProposerSlashingsSchema() {
+    return (SszListSchema<ProposerSlashing, ?>) getFieldSchema3();
+  }
+
+  @SuppressWarnings("unchecked")
+  @Override
+  public SszListSchema<AttesterSlashing, ?> getAttesterSlashingsSchema() {
+    return (SszListSchema<AttesterSlashing, ?>) getFieldSchema4();
+  }
+
+  @SuppressWarnings("unchecked")
+  @Override
+  public SszListSchema<Attestation, ?> getAttestationsSchema() {
+    return (SszListSchema<Attestation, ?>) getFieldSchema5();
+  }
+
+  @SuppressWarnings("unchecked")
+  @Override
+  public SszListSchema<Deposit, ?> getDepositsSchema() {
+    return (SszListSchema<Deposit, ?>) getFieldSchema6();
+  }
+
+  @SuppressWarnings("unchecked")
+  @Override
+  public SszListSchema<SignedVoluntaryExit, ?> getVoluntaryExitsSchema() {
+    return (SszListSchema<SignedVoluntaryExit, ?>) getFieldSchema7();
+  }
+
+  @Override
+  public SyncAggregateSchema getSyncAggregateSchema() {
+    return (SyncAggregateSchema) getFieldSchema8();
+  }
+
+  @Override
+  public BeaconBlockBodyAltairImpl createFromBackingNode(TreeNode node) {
+    return new BeaconBlockBodyAltairImpl(this, node);
+  }
+}
