+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/blocks/blockbody/versions/merge/BeaconBlockBodyMerge.java
@@ -16,71 +16,14 @@ package tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.merge;
 import static com.google.common.base.Preconditions.checkArgument;
 
 import java.util.Optional;
-import org.apache.tuweni.bytes.Bytes32;
-import tech.pegasys.teku.bls.BLSSignature;
-import tech.pegasys.teku.spec.datastructures.blocks.Eth1Data;
 import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBody;
 import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.altair.BeaconBlockBodyAltair;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayload;
-import tech.pegasys.teku.spec.datastructures.operations.Attestation;
-import tech.pegasys.teku.spec.datastructures.operations.AttesterSlashing;
-import tech.pegasys.teku.spec.datastructures.operations.Deposit;
-import tech.pegasys.teku.spec.datastructures.operations.ProposerSlashing;
-import tech.pegasys.teku.spec.datastructures.operations.SignedVoluntaryExit;
-import tech.pegasys.teku.spec.datastructures.type.SszSignature;
-import tech.pegasys.teku.ssz.SszList;
-import tech.pegasys.teku.ssz.containers.Container9;
-import tech.pegasys.teku.ssz.primitive.SszBytes32;
-import tech.pegasys.teku.ssz.tree.TreeNode;
 
 /** A Beacon block body */
-public class BeaconBlockBodyMerge
-    extends Container9<
-        BeaconBlockBodyMerge,
-        SszSignature,
-        Eth1Data,
-        SszBytes32,
-        SszList<ProposerSlashing>,
-        SszList<AttesterSlashing>,
-        SszList<Attestation>,
-        SszList<Deposit>,
-        SszList<SignedVoluntaryExit>,
-        ExecutionPayload>
-    implements BeaconBlockBody {
+public interface BeaconBlockBodyMerge extends BeaconBlockBodyAltair {
 
-  BeaconBlockBodyMerge(BeaconBlockBodySchemaMerge type) {
-    super(type);
-  }
-
-  BeaconBlockBodyMerge(BeaconBlockBodySchemaMerge type, TreeNode backingNode) {
-    super(type, backingNode);
-  }
-
-  BeaconBlockBodyMerge(
-      BeaconBlockBodySchemaMerge type,
-      SszSignature randao_reveal,
-      Eth1Data eth1_data,
-      SszBytes32 graffiti,
-      SszList<ProposerSlashing> proposer_slashings,
-      SszList<AttesterSlashing> attester_slashings,
-      SszList<Attestation> attestations,
-      SszList<Deposit> deposits,
-      SszList<SignedVoluntaryExit> voluntary_exits,
-      ExecutionPayload execution_payload) {
-    super(
-        type,
-        randao_reveal,
-        eth1_data,
-        graffiti,
-        proposer_slashings,
-        attester_slashings,
-        attestations,
-        deposits,
-        voluntary_exits,
-        execution_payload);
-  }
-
-  public static BeaconBlockBodyMerge required(final BeaconBlockBody body) {
+  static BeaconBlockBodyMerge required(final BeaconBlockBody body) {
     checkArgument(
         body instanceof BeaconBlockBodyMerge,
         "Expected merge block body but got %s",
@@ -88,62 +31,13 @@ public class BeaconBlockBodyMerge
     return (BeaconBlockBodyMerge) body;
   }
 
-  @Override
-  public BLSSignature getRandao_reveal() {
-    return getField0().getSignature();
-  }
+  ExecutionPayload getExecution_payload();
 
   @Override
-  public Eth1Data getEth1_data() {
-    return getField1();
-  }
+  BeaconBlockBodySchemaMerge<?> getSchema();
 
   @Override
-  public Bytes32 getGraffiti() {
-    return getField2().get();
-  }
-
-  @Override
-  public SszList<ProposerSlashing> getProposer_slashings() {
-    return getField3();
-  }
-
-  @Override
-  public SszList<AttesterSlashing> getAttester_slashings() {
-    return getField4();
-  }
-
-  @Override
-  public SszList<Attestation> getAttestations() {
-    return getField5();
-  }
-
-  @Override
-  public SszList<Deposit> getDeposits() {
-    return getField6();
-  }
-
-  @Override
-  public SszList<SignedVoluntaryExit> getVoluntary_exits() {
-    return getField7();
-  }
-
-  public ExecutionPayload getExecution_payload() {
-    return getField8();
-  }
-
-  @Override
-  public BeaconBlockBodySchemaMerge getSchema() {
-    return (BeaconBlockBodySchemaMerge) super.getSchema();
-  }
-
-  @Override
-  public Optional<BeaconBlockBodyMerge> toVersionMerge() {
+  default Optional<BeaconBlockBodyMerge> toVersionMerge() {
     return Optional.of(this);
-  }
-
-  @Override
-  public Optional<BeaconBlockBodyAltair> toVersionAltair() {
-    return Optional.empty();
   }
 }
