+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/blocks/blockbody/versions/merge/BeaconBlockBodySchemaMerge.java
@@ -14,154 +14,29 @@
 package tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.merge;
 
 import java.util.Optional;
-import java.util.function.Consumer;
 import tech.pegasys.teku.spec.config.SpecConfig;
 import tech.pegasys.teku.spec.config.SpecConfigMerge;
-import tech.pegasys.teku.spec.datastructures.blocks.Eth1Data;
-import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBodyBuilder;
-import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBodySchema;
-import tech.pegasys.teku.spec.datastructures.blocks.blockbody.common.BlockBodyFields;
-import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayload;
-import tech.pegasys.teku.spec.datastructures.operations.Attestation;
-import tech.pegasys.teku.spec.datastructures.operations.AttesterSlashing;
-import tech.pegasys.teku.spec.datastructures.operations.Deposit;
-import tech.pegasys.teku.spec.datastructures.operations.ProposerSlashing;
-import tech.pegasys.teku.spec.datastructures.operations.SignedVoluntaryExit;
-import tech.pegasys.teku.spec.datastructures.type.SszSignature;
-import tech.pegasys.teku.spec.datastructures.type.SszSignatureSchema;
-import tech.pegasys.teku.ssz.SszList;
-import tech.pegasys.teku.ssz.containers.ContainerSchema9;
-import tech.pegasys.teku.ssz.primitive.SszBytes32;
-import tech.pegasys.teku.ssz.schema.SszListSchema;
-import tech.pegasys.teku.ssz.schema.SszPrimitiveSchemas;
-import tech.pegasys.teku.ssz.tree.TreeNode;
+import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.altair.BeaconBlockBodySchemaAltair;
 
-public class BeaconBlockBodySchemaMerge
-    extends ContainerSchema9<
-        BeaconBlockBodyMerge,
-        SszSignature,
-        Eth1Data,
-        SszBytes32,
-        SszList<ProposerSlashing>,
-        SszList<AttesterSlashing>,
-        SszList<Attestation>,
-        SszList<Deposit>,
-        SszList<SignedVoluntaryExit>,
-        ExecutionPayload>
-    implements BeaconBlockBodySchema<BeaconBlockBodyMerge> {
+public interface BeaconBlockBodySchemaMerge<T extends BeaconBlockBodyMerge>
+    extends BeaconBlockBodySchemaAltair<T> {
 
-  private BeaconBlockBodySchemaMerge(
-      NamedSchema<SszSignature> randaoRevealSchema,
-      NamedSchema<Eth1Data> eth1DataSchema,
-      NamedSchema<SszBytes32> graffitiSchema,
-      NamedSchema<SszList<ProposerSlashing>> proposerSlashingsSchema,
-      NamedSchema<SszList<AttesterSlashing>> attesterSlashingsSchema,
-      NamedSchema<SszList<Attestation>> attestationsSchema,
-      NamedSchema<SszList<Deposit>> depositsSchema,
-      NamedSchema<SszList<SignedVoluntaryExit>> voluntaryExitsSchema,
-      NamedSchema<ExecutionPayload> executionPayloadSchema) {
-    super(
-        "BeaconBlockBody",
-        randaoRevealSchema,
-        eth1DataSchema,
-        graffitiSchema,
-        proposerSlashingsSchema,
-        attesterSlashingsSchema,
-        attestationsSchema,
-        depositsSchema,
-        voluntaryExitsSchema,
-        executionPayloadSchema);
-  }
-
-  public static BeaconBlockBodySchemaMerge create(final SpecConfig specConfig) {
+  static BeaconBlockBodySchemaMerge<?> create(final SpecConfig specConfig) {
     return SpecConfigMerge.required(
         specConfig,
         config ->
-            create(
+            BeaconBlockBodySchemaMergeImpl.create(
                 config.getMaxProposerSlashings(),
                 config.getMaxAttesterSlashings(),
                 config.getMaxAttestations(),
                 config.getMaxDeposits(),
-                config.getMaxVoluntaryExits()));
-  }
-
-  private static BeaconBlockBodySchemaMerge create(
-      final long maxProposerSlashings,
-      final long maxAttesterSlashings,
-      final long maxAttestations,
-      final long maxDeposits,
-      final long maxVoluntaryExits) {
-    return new BeaconBlockBodySchemaMerge(
-        namedSchema(BlockBodyFields.RANDAO_REVEAL.name(), SszSignatureSchema.INSTANCE),
-        namedSchema(BlockBodyFields.ETH1_DATA.name(), Eth1Data.SSZ_SCHEMA),
-        namedSchema(BlockBodyFields.GRAFFITI.name(), SszPrimitiveSchemas.BYTES32_SCHEMA),
-        namedSchema(
-            BlockBodyFields.PROPOSER_SLASHINGS.name(),
-            SszListSchema.create(ProposerSlashing.SSZ_SCHEMA, maxProposerSlashings)),
-        namedSchema(
-            BlockBodyFields.ATTESTER_SLASHINGS.name(),
-            SszListSchema.create(AttesterSlashing.SSZ_SCHEMA, maxAttesterSlashings)),
-        namedSchema(
-            BlockBodyFields.ATTESTATIONS.name(),
-            SszListSchema.create(Attestation.SSZ_SCHEMA, maxAttestations)),
-        namedSchema(
-            BlockBodyFields.DEPOSITS.name(), SszListSchema.create(Deposit.SSZ_SCHEMA, maxDeposits)),
-        namedSchema(
-            BlockBodyFields.VOLUNTARY_EXITS.name(),
-            SszListSchema.create(SignedVoluntaryExit.SSZ_SCHEMA, maxVoluntaryExits)),
-        namedSchema(BlockBodyFields.EXECUTION_PAYLOAD.name(), ExecutionPayload.SSZ_SCHEMA));
+                config.getMaxVoluntaryExits(),
+                // TODO SpecConfigMerge inherits SpecConfigAltair
+                config.toVersionAltair().orElseThrow().getSyncCommitteeSize()));
   }
 
   @Override
-  public BeaconBlockBodyMerge createBlockBody(
-      final Consumer<BeaconBlockBodyBuilder> builderConsumer) {
-    final BeaconBlockBodyBuilderMerge builder = new BeaconBlockBodyBuilderMerge().schema(this);
-    builderConsumer.accept(builder);
-    return builder.build();
-  }
-
-  @Override
-  public BeaconBlockBodyMerge createEmpty() {
-    return new BeaconBlockBodyMerge(this);
-  }
-
-  @SuppressWarnings("unchecked")
-  @Override
-  public SszListSchema<ProposerSlashing, ?> getProposerSlashingsSchema() {
-    return (SszListSchema<ProposerSlashing, ?>) getFieldSchema3();
-  }
-
-  @SuppressWarnings("unchecked")
-  @Override
-  public SszListSchema<AttesterSlashing, ?> getAttesterSlashingsSchema() {
-    return (SszListSchema<AttesterSlashing, ?>) getFieldSchema4();
-  }
-
-  @SuppressWarnings("unchecked")
-  @Override
-  public SszListSchema<Attestation, ?> getAttestationsSchema() {
-    return (SszListSchema<Attestation, ?>) getFieldSchema5();
-  }
-
-  @SuppressWarnings("unchecked")
-  @Override
-  public SszListSchema<Deposit, ?> getDepositsSchema() {
-    return (SszListSchema<Deposit, ?>) getFieldSchema6();
-  }
-
-  @SuppressWarnings("unchecked")
-  @Override
-  public SszListSchema<SignedVoluntaryExit, ?> getVoluntaryExitsSchema() {
-    return (SszListSchema<SignedVoluntaryExit, ?>) getFieldSchema7();
-  }
-
-  @Override
-  public BeaconBlockBodyMerge createFromBackingNode(TreeNode node) {
-    return new BeaconBlockBodyMerge(this, node);
-  }
-
-  @Override
-  public Optional<BeaconBlockBodySchemaMerge> toVersionMerge() {
+  default Optional<BeaconBlockBodySchemaMerge<?>> toVersionMerge() {
     return Optional.of(this);
   }
 }
